Change the default way to report test status
The previous way using selenium was not scaling correctly to large test
suites (300 tests). IE7 was really slow and IE10 was reporting invalid
data.
We now post a form which is much faster and reliable (it was already
used for coverage data).

Note: The API of the onTest method of the Remote class  was changed
in a non-backward compatible way.